\name{eyetracking-package}
\alias{eyetracking}
\docType{package}
\title{
Eyetracking Helper Functions
}
\description{
Misc helper functions for working with eyetracking data.
}
\details{
\tabular{ll}{
Package: \tab eyetracking\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2011-05-25\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Ryan M. Hope

Maintainer: Ryan M. Hope <rmh3093@gmail.com>
}

\keyword{ eyetracking }
